requirements.txt
fastapi
langchain
langchain-community
langchain-core
langchain-openai
langchain_experimental
langsmith
openai
pydantic
pandas
pydantic_core
python-dotenv
python-multipart
uvicorn
tabulate
openpyxl
streamlit
plotly